Update main.c
Application Reverts To only Using Slot 0 When No Multi-tap is Inserted. 
Resolves The 4x issue when no multi-tap is connected.

Added in On Screen Indicators To the Check/Format Function to Let the User Know If a Multi-tap is not present when Checking the Status or Formatting. 

To Summarize, These Changes Essentially cause the app to Revert to the Original Application Code when No multi-tap is present This will Prevent the Application From Formatting or Getting the information 4x on a Memorycard Connected to a port with no multi-tap